+++ hnix-store-core/src/System/Nix/Build.hs
@@ -1,5 +1,3 @@
--- due to recent generic-arbitrary
-{-# OPTIONS_GHC -fconstraint-solver-iterations=0 #-}
 {-|
 Description : Build related types
 Maintainer  : srk <srk@48.io>
@@ -14,15 +12,11 @@ module System.Nix.Build
 import Data.Time (UTCTime)
 import Data.Text (Text)
 import GHC.Generics (Generic)
-import Test.QuickCheck (Arbitrary(..))
-import Test.QuickCheck.Arbitrary.Generic (GenericArbitrary(..))
-import Test.QuickCheck.Instances ()
 
 -- keep the order of these Enums to match enums from reference implementations
 -- src/libstore/store-api.hh
 data BuildMode = Normal | Repair | Check
   deriving (Eq, Generic, Ord, Enum, Show)
-  deriving Arbitrary via GenericArbitrary BuildMode
 
 data BuildStatus =
     Built
@@ -41,7 +35,6 @@ data BuildStatus =
   | ResolvesToAlreadyValid
   | NoSubstituters
   deriving (Eq, Generic, Ord, Enum, Show)
-  deriving Arbitrary via GenericArbitrary BuildStatus
 
 -- | Result of the build
 data BuildResult = BuildResult
@@ -59,7 +52,6 @@ data BuildResult = BuildResult
     stopTime           :: !UTCTime
   }
   deriving (Eq, Generic, Ord, Show)
-  deriving Arbitrary via GenericArbitrary BuildResult
 
 buildSuccess :: BuildResult -> Bool
 buildSuccess BuildResult {..} =
