+++ hnix-store-core/tests/ContentAddress.hs
@@ -1,18 +1,21 @@
-module ContentAddress where
+module ContentAddressSpec where
 
-import Test.Tasty.QuickCheck
-import System.Nix.ContentAddress (ContentAddress)
+import Test.Hspec (Spec, describe, shouldBe)
+import Test.Hspec.QuickCheck (prop)
+import System.Nix.Arbitrary ()
 
 import qualified Data.Attoparsec.Text.Lazy
 import qualified Data.Text.Lazy.Builder
 
 import qualified System.Nix.ContentAddress
 
-prop_caAddrRoundTrip :: ContentAddress -> Property
-prop_caAddrRoundTrip = \caAddr ->
-  Data.Attoparsec.Text.Lazy.parseOnly
-    System.Nix.ContentAddress.contentAddressParser
-      (Data.Text.Lazy.Builder.toLazyText
-        (System.Nix.ContentAddress.contentAddressBuilder caAddr))
-  === pure caAddr
+spec :: Spec
+spec = do
+  describe "ContentAddress" $ do
+    prop "roundtrips" $ \caAddr ->
+      Data.Attoparsec.Text.Lazy.parseOnly
+        System.Nix.ContentAddress.contentAddressParser
+          (Data.Text.Lazy.Builder.toLazyText
+            (System.Nix.ContentAddress.contentAddressBuilder caAddr))
+      `shouldBe` pure caAddr
