+++ hnix-store-remote/hnix-store-remote.cabal
@@ -116,6 +116,7 @@ test-suite remote
       base >=4.12 && <5
     , hnix-store-core
     , hnix-store-remote
+    , hnix-store-tests
     , bytestring
     , cereal
     , text
