+++ hnix-store-tests/CHANGELOG.md
@@ -1,0 +1,10 @@
+# Version [0.1.0.0](https://github.com/haskell-nix/hnix-store/compare/tests-0.1.0.0...tests-0.1.1.0) (2023-11-27)
+
+* Initial release
+
+---
+
+`hnix-store-tests` uses [PVP Versioning][1].
+
+[1]: https://pvp.haskell.org
+
