+++ hnix-store-tests/README.md
@@ -1,0 +1,7 @@
+# hnix-store-tests
+
+Arbitrary instances for core types, roundtrip property tests, utility functions
+and a harness for running tests that require nix-store
+with nix-daemon.
+
+[Test.Hspec.Nix]: ./src/Test/Hspec/Nix.hs
