+++ hnix-store-tests/hnix-store-tests.cabal
@@ -1,0 +1,81 @@
+cabal-version:       2.2
+name:                hnix-store-tests
+version:             0.1.0.0
+synopsis:            Test utilities and instances
+description:
+  This package contains Arbitrary instances for core
+  types, roundtrip property tests, utility functions
+  and a harness for running tests that require nix-store
+  with nix-daemon.
+homepage:            https://github.com/haskell-nix/hnix-store
+license:             Apache-2.0
+license-file:        LICENSE
+author:              Richard Marko
+maintainer:          srk@48.io
+copyright:           2023 Richard Marko
+category:            Nix
+build-type:          Simple
+extra-source-files:
+    CHANGELOG.md
+  , README.md
+
+common commons
+  ghc-options:  -Wall
+  default-extensions:
+      DerivingStrategies
+    , DerivingVia
+    , FlexibleInstances
+    , ScopedTypeVariables
+    , StandaloneDeriving
+    , RecordWildCards
+    , TypeApplications
+    , LambdaCase
+  default-language: Haskell2010
+
+library
+  import: commons
+  exposed-modules:
+      System.Nix.Arbitrary
+    , System.Nix.Arbitrary.Build
+    , System.Nix.Arbitrary.ContentAddress
+    , System.Nix.Arbitrary.Derivation
+    , System.Nix.Arbitrary.DerivedPath
+    , System.Nix.Arbitrary.Hash
+    , System.Nix.Arbitrary.StorePath
+  build-depends:
+      base >=4.12 && <5
+    , hnix-store-core
+    , bytestring
+    , cryptonite
+    , dependent-sum > 0.7
+    , generic-arbitrary < 1.1
+    , QuickCheck
+    , quickcheck-instances
+    , text
+  hs-source-dirs:      src
+
+test-suite props
+  import: commons
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  other-modules:
+      BaseEncodingSpec
+      ContentAddressSpec
+      DerivationSpec
+      DerivedPathSpec
+      StorePathSpec
+  hs-source-dirs:
+    tests
+  build-tool-depends:
+    hspec-discover:hspec-discover
+  build-depends:
+      base
+    , hnix-store-core
+    , hnix-store-tests
+    , attoparsec
+    , bytestring
+    , containers
+    , data-default-class
+    , QuickCheck
+    , text
+    , hspec
