+++ hnix-store-tests/src/System/Nix/Arbitrary.hs
@@ -1,0 +1,8 @@
+module System.Nix.Arbitrary where
+
+import System.Nix.Arbitrary.Build ()
+import System.Nix.Arbitrary.ContentAddress ()
+import System.Nix.Arbitrary.Derivation ()
+import System.Nix.Arbitrary.DerivedPath ()
+import System.Nix.Arbitrary.Hash ()
+import System.Nix.Arbitrary.StorePath ()
